requirements.txt
openpyxl==3.0.10
plotly==4.14.3
pandas==1.4.3
streamlit==1.12.0
Pillow==9.2.0
tabulate==0.9.0
tensorflow==2.13.0
streamlit-aggrid==0.3.4.post3
streamlit-lottie==0.0.5
matplotlib==3.4.3
scikit-learn==1.0.1
seaborn==0.12.2
keras==2.13.1
numpy==1.24.3



click==8.0
